requirements.txt
flask
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
psycopg2-binary
ebooklib
beautifulsoup4